:green_heart: CI compliant seeders

+++ database/seeders/AppSettingsSeeder.php
@@ -5,10 +5,9 @@ namespace Database\Seeders;
 use App\Enums\AppTheme;
 use App\Models\AppSettings;
 use App\Services\AppSettingsManager;
-use Illuminate\Database\Seeder;
 use Throwable;
 
-class AppSettingsSeeder extends Seeder
+class AppSettingsSeeder extends CiCdCompliantSeeder
 {
     /**
      * Run the database seeds.
@@ -17,6 +16,10 @@ class AppSettingsSeeder extends Seeder
      */
     public function run(): void
     {
+        if ($this->tableNotEmpty()) {
+            return;
+        }
+
         AppSettings::query()->delete();
         $settingsManager = resolve(AppSettingsManager::class);
 
@@ -26,5 +29,10 @@ class AppSettingsSeeder extends Seeder
         ];
 
         $settingsManager->setSettings($default);
+    }
+
+    protected function tableName(): string
+    {
+        return app(AppSettings::class)->getTable();
     }
 }
